Created account module, account details component so its reuseable across registration, account creation and within changing account details (#929)
